Move helper function to LSPFileUpdates
