+++ main/lsp/LSPTypechecker.cc
@@ -41,33 +41,6 @@ void sendTypecheckInfo(const LSPConfiguration &config, const core::GlobalState &
     }
 }
 
-// In debug builds, asserts that we have not accidentally taken the fast path after a change to the set of
-// methods in a file.
-bool validateIdenticalFingerprints(const std::vector<core::SymbolHash> &a, const std::vector<core::SymbolHash> &b) {
-    if (a.size() != b.size()) {
-        return false;
-    }
-
-    core::SymbolHash previousHash; // Initializes to <0, 0>.
-    auto bIt = b.begin();
-    for (const auto &methodA : a) {
-        const auto &methodB = *bIt;
-        if (methodA.nameHash != methodB.nameHash) {
-            return false;
-        }
-
-        // Enforce that hashes are sorted in ascending order.
-        if (methodA < previousHash) {
-            return false;
-        }
-
-        previousHash = methodA;
-        bIt++;
-    }
-
-    return true;
-}
-
 vector<ast::ParsedFile> sortParsedFiles(const core::GlobalState &gs, ErrorReporter &errorReporter,
                                         vector<ast::ParsedFile> parsedFiles) {
     fast_sort(parsedFiles, [&](const auto &lhs, const auto &rhs) -> bool {
@@ -221,126 +194,6 @@ bool LSPTypechecker::typecheck(LSPFileUpdates updates, WorkerPool &workers,
     return committed;
 }
 
-namespace {
-
-struct FastPathFilesToTypecheckResult {
-    // size_t is an index into the LSPFileUpdates::updatedFiles vector
-    UnorderedMap<core::FileRef, size_t> changedFiles;
-
-    // The names of all symbols changed by this set of updates
-    vector<core::ShortNameHash> changedSymbolNameHashes;
-
-    // Extra files that need to be typechecked because the file mentions the name of one of the changed symbols.
-    vector<core::FileRef> extraFiles;
-};
-
-FastPathFilesToTypecheckResult getFilesToTypecheck(const core::GlobalState &gs, const LSPConfiguration &config,
-                                                   const vector<shared_ptr<core::File>> &updatedFiles) {
-    FastPathFilesToTypecheckResult result;
-    Timer timeit(config.logger, "compute_fast_path_file_set");
-    vector<core::SymbolHash> changedMethodSymbolHashes;
-    vector<core::SymbolHash> changedFieldSymbolHashes;
-    auto idx = -1;
-    for (const auto &updatedFile : updatedFiles) {
-        idx++;
-        auto fref = gs.findFileByPath(updatedFile->path());
-        // We don't support new files on the fast path. This enforce failing indicates a bug in our fast/slow
-        // path logic in LSPPreprocessor.
-        ENFORCE(fref.exists());
-        ENFORCE(updatedFile->getFileHash() != nullptr);
-        if (config.opts.stripePackages && updatedFile->isPackage()) {
-            // Only relevant in --stripe-packages mode. Package declarations do not have method
-            // hashes. Instead we rely on recomputing packages if any __package.rb source
-            // changes.
-            continue;
-        }
-        if (fref.exists()) {
-            // Update to existing file on fast path
-            ENFORCE(fref.data(gs).getFileHash() != nullptr);
-            const auto &oldSymbolHashes = fref.data(gs).getFileHash()->localSymbolTableHashes;
-            const auto &newSymbolHashes = updatedFile->getFileHash()->localSymbolTableHashes;
-            const auto &oldMethodHashes = oldSymbolHashes.methodHashes;
-            const auto &newMethodHashes = newSymbolHashes.methodHashes;
-
-            if (config.opts.lspExperimentalFastPathEnabled) {
-                // Find which hashes changed. Note: methodHashes are sorted, so set_difference should work.
-                // This will insert two entries into `changedMethodHashes` for each changed method, but they
-                // will get deduped later.
-                absl::c_set_symmetric_difference(oldMethodHashes, newMethodHashes,
-                                                 std::back_inserter(changedMethodSymbolHashes));
-            } else {
-                // Both oldHash and newHash should have the same methods, since this is the fast path!
-                ENFORCE(validateIdenticalFingerprints(oldMethodHashes, newMethodHashes),
-                        "definitionHash should have failed");
-
-                // Find which hashes changed. Note: methodHashes are sorted, so set_difference should work.
-                // This will insert two entries into `changedMethodHashes` for each changed method, but they
-                // will get deduped later.
-                absl::c_set_difference(oldMethodHashes, newMethodHashes, std::back_inserter(changedMethodSymbolHashes));
-            }
-
-            const auto &oldFieldHashes = oldSymbolHashes.staticFieldHashes;
-            const auto &newFieldHashes = newSymbolHashes.staticFieldHashes;
-
-            ENFORCE(validateIdenticalFingerprints(oldFieldHashes, newFieldHashes), "definitionHash should have failed");
-
-            absl::c_set_difference(oldFieldHashes, newFieldHashes, std::back_inserter(changedFieldSymbolHashes));
-
-            result.changedFiles.emplace(fref, idx);
-        }
-    }
-
-    result.changedSymbolNameHashes.reserve(changedMethodSymbolHashes.size() + changedFieldSymbolHashes.size());
-    absl::c_transform(changedMethodSymbolHashes, std::back_inserter(result.changedSymbolNameHashes),
-                      [](const auto &symhash) { return symhash.nameHash; });
-    absl::c_transform(changedFieldSymbolHashes, std::back_inserter(result.changedSymbolNameHashes),
-                      [](const auto &symhash) { return symhash.nameHash; });
-    core::ShortNameHash::sortAndDedupe(result.changedSymbolNameHashes);
-
-    if (!result.changedSymbolNameHashes.empty()) {
-        // ^ optimization--skip the loop over every file in the project (`gs.getFiles()`) if
-        // the set of changed symbols is empty (e.g., running a completion request inside a
-        // method body)
-        int i = -1;
-        for (auto &oldFile : gs.getFiles()) {
-            i++;
-            if (oldFile == nullptr) {
-                continue;
-            }
-
-            auto ref = core::FileRef(i);
-            if (result.changedFiles.contains(ref)) {
-                continue;
-            }
-
-            if (config.opts.stripePackages && oldFile->isPackage()) {
-                continue; // See note above about --stripe-packages.
-            }
-
-            if (oldFile->isPayload()) {
-                // Don't retypecheck files in the payload via incremental namer, as that might
-                // cause well-known symbols to get deleted and assigned a new SymbolRef ID.
-                continue;
-            }
-
-            ENFORCE(oldFile->getFileHash() != nullptr);
-            const auto &oldHash = *oldFile->getFileHash();
-            vector<core::ShortNameHash> intersection;
-            absl::c_set_intersection(result.changedSymbolNameHashes, oldHash.usages.nameHashes,
-                                     std::back_inserter(intersection));
-            if (intersection.empty()) {
-                continue;
-            }
-
-            result.extraFiles.emplace_back(ref);
-        }
-    }
-
-    return result;
-}
-
-} // namespace
-
 vector<core::FileRef> LSPTypechecker::runFastPath(LSPFileUpdates &updates, WorkerPool &workers,
                                                   shared_ptr<core::ErrorFlusher> errorFlusher) const {
     ENFORCE(this_thread::get_id() == typecheckerThreadId, "Typechecker can only be used from the typechecker thread.");
@@ -357,7 +210,7 @@ vector<core::FileRef> LSPTypechecker::runFastPath(LSPFileUpdates &updates, Worke
     Timer timeit(config->logger, "fast_path");
     // Replace error queue with one that is owned by this thread.
     gs->errorQueue = make_shared<core::ErrorQueue>(gs->errorQueue->logger, gs->errorQueue->tracer, errorFlusher);
-    auto result = getFilesToTypecheck(*gs, *config, updates.updatedFiles);
+    auto result = updates.fastPathFilesToTypecheck(*gs, *config);
     config->logger->debug("Added {} files that were not part of the edit to the update set", result.extraFiles.size());
     UnorderedMap<core::FileRef, core::FoundMethodHashes> oldFoundMethodHashesForFiles;
     auto toTypecheck = move(result.extraFiles);
